+++ index.js
@@ -111,7 +111,7 @@ function compare (a, b) {
   return crypto.timingSafeEqual(a, b)
 }
 
-function plugin (fastify, options, done) {
+function fastifyBearerAuth (fastify, options, done) {
   const defaultLogLevel = 'error'
   options = Object.assign({ addHook: true, verifyErrorLogLevel: defaultLogLevel }, options)
 
@@ -136,8 +136,11 @@ function plugin (fastify, options, done) {
   done()
 }
 
-module.exports = fp(plugin, {
+module.exports = fp(fastifyBearerAuth, {
   fastify: '4.x',
   name: '@fastify/bearer-auth'
 })
+module.exports.default = fastifyBearerAuth
+module.exports.fastifyBearerAuth = fastifyBearerAuth
+
 module.exports.internals = { factory: verifyBearerAuthFactory }
